add mooneye tests
run all mooneye acceptance tests. takes about 30 seconds. 168 pass, 313 fail, and 1 errored.

filter tests for unsupported hardware

fix bits/mem_oam test

rearrange clock increment and video update

remove symbol files

move tests that the tcrs emulator also failed

This way I am only testing my work porting, not also layering in issues with the initial emulator.

The original was able to play first gen pokemon, that is my goal. No need to go super deep into accuracy at this time. And that may actually be detrimental to going fast anyway.

mooneye tests now only report two values instead of 7

previously each test rom would use @test to check completion and once per register. This inflated the numbers and added noise to each test run.

Now that the tests are collapsed into one statement with boolean logic the output is faster with less scrollback and the numbers are easier to visually parse.

oops. single negation causes all timer tests to pass or fail

one explicit type conversion down, two more test cases to go.

typos

banked ram write

ignore traces

fix timing for jr cc e instruction

+++ src/Carts.jl
@@ -40,7 +40,7 @@ function Component.write!(mbc::MBC1, addr::UInt16, data::UInt8, ram::Ref{OffsetV
         mbc.active_ram_bank = data & 0x03
     elseif 0xa000 <= addr < 0xc000
         reladdr = addr - 0xa0000
-        ram[][mbc.active_ram_bank * 0x2000 + reladdr] = data
+        ram[][UInt16(mbc.active_ram_bank) * 0x2000 + reladdr] = data
     end
     nothing
 end
